+++ CHANGELOG.rst
@@ -36,6 +36,8 @@ Fixed
 ^^^^^
 - ``ActionParser`` not updating ``dest`` of groups for instantiation (`#707
   <https://github.com/omni-us/jsonargparse/pull/707>`__).
+- Nested generic dataclasses not working correctly (`#709
+  <https://github.com/omni-us/jsonargparse/pull/709>`__).
 
 
 v4.38.0 (2025-03-26)